Any kind of Editext can be a PatternedEdittext using PatternedViewHelper base functions. Default PatternedEdittext behaviour remains the same.
